+++ server/src/main/java/com/exacaster/lighter/application/sessions/SessionService.java
@@ -8,7 +8,7 @@ import com.exacaster.lighter.application.SubmitParams;
 import com.exacaster.lighter.application.sessions.exceptions.InvalidSessionStateException;
 import com.exacaster.lighter.application.sessions.processors.StatementHandler;
 import com.exacaster.lighter.backend.Backend;
-import com.exacaster.lighter.rest.SessionParameters;
+import com.exacaster.lighter.rest.SessionParams;
 import com.exacaster.lighter.storage.ApplicationStorage;
 import com.exacaster.lighter.storage.SortOrder;
 import com.exacaster.lighter.storage.StatementStorage;
@@ -47,15 +47,15 @@ public class SessionService {
         return applicationStorage.findApplications(ApplicationType.SESSION, from, size);
     }
 
-    public Application createSession(SessionParameters sessionParameters) {
-        return createSession(UUID.randomUUID().toString(), sessionParameters);
+    public Application createSession(SessionParams sessionParams) {
+        return createSession(UUID.randomUUID().toString(), sessionParams);
     }
 
-    public Application createSession(String sessionId, SessionParameters sessionParameters) {
-        if(Boolean.TRUE.equals(sessionParameters.getPermanent())){
-            return createPermanentSession(sessionId, sessionParameters);
+    public Application createSession(String sessionId, SessionParams sessionParams) {
+        if(Boolean.TRUE.equals(sessionParams.getPermanent())){
+            return createPermanentSession(sessionId, sessionParams);
         }
-        return createRegularSession(sessionId, sessionParameters);
+        return createRegularSession(sessionId, sessionParams);
     }
 
     public Application createPermanentSession(String sessionId, SubmitParams params) {
